use int64/wxLongLong in Model_Translink and json_writer.Int64(xxx.GetValue())

+++ src/model/Model_Translink.cpp
@@ -50,7 +50,7 @@ Model_Translink& Model_Translink::instance()
     return Singleton<Model_Translink>::instance();
 }
 
-Model_Translink::CHECKING_TYPE Model_Translink::type_checking(const int tt)
+Model_Translink::CHECKING_TYPE Model_Translink::type_checking(const int64 tt)
 {
     if (tt == AS_TRANSFER)
     {
